requirements.txt
Django==6.0.1
djangorestframework==3.16.1
django-cors-headers==4.4.0
firebase-admin
sentry-sdk
gunicorn
dj-database-url
whitenoise
psycopg2-binary
